Split project repositories into a rule
