+++ subprojects/gradle-plugin-development/src/main/java/dev/gradleplugins/internal/plugins/GradlePluginDevelopmentExtensionPlugin.java
@@ -6,7 +6,7 @@ import org.gradle.api.Project;
 public abstract class GradlePluginDevelopmentExtensionPlugin implements Plugin<Project> {
     @Override
     public void apply(Project project) {
-        new RegisterGradlePluginDevelopmentRepositoryExtensionRule().execute(project);
+        project.getPluginManager().apply("gradlepluginsdev.rules.project-repositories-extension");
         project.getPluginManager().apply("gradlepluginsdev.rules.project-dependencies-extension");
     }
 }
