+++ subprojects/gradle-plugin-development/src/main/resources/META-INF/gradle-plugins/gradlepluginsdev.rules.settings-repositories-extension.properties
@@ -1,0 +1,1 @@
+implementation-class=dev.gradleplugins.internal.rules.RepositoriesExtensionRules$ForSettings
