+++ subprojects/gradle-plugin-development/src/test/groovy/dev/gradleplugins/GradlePluginDevelopmentRepositoryExtensionTest.java
@@ -3,6 +3,7 @@ package dev.gradleplugins;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository;
 import org.gradle.testfixtures.ProjectBuilder;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,7 +11,13 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 class GradlePluginDevelopmentRepositoryExtensionTest {
     private final Project project = ProjectBuilder.builder().build();
-    private final GradlePluginDevelopmentRepositoryExtension subject = GradlePluginDevelopmentRepositoryExtension.from(project.getRepositories());
+    private GradlePluginDevelopmentRepositoryExtension subject;
+
+    @BeforeEach
+    void setup() {
+        project.getPluginManager().apply("dev.gradleplugins.gradle-plugin-base");
+        subject = GradlePluginDevelopmentRepositoryExtension.from(project.getRepositories());
+    }
 
     @Test
     void returnsGradlePluginDevelopmentRepository() {
